feat: add local test environment scripts and database handling
- Introduced `local_test_env.sh` for loading environment variables and managing database URLs for local testing.
- Created `run_e2e_regression_tests.sh` to execute end-to-end regression tests against a local Supabase/Postgres setup.
- Added `run_me_endpoint_tests.sh` for running integration tests on the /me endpoint with local database support.
- Implemented `run_search_api_tests.sh` to run search API tests in both default and database-backed modes, with enhanced error handling for database connectivity.
- Developed `scan_frontend_lib_i18n.py` to scan Dart files for user-facing string literals, categorizing them for localization efforts.
- Added `supabase_start_db_local.sh` to start only the necessary local Supabase Postgres service for testing, avoiding conflicts with other services.
- Created a new migration file to add a `suggested_action` column to the `app_quality_review` table for improved quality review processes.

+++ backend/src/billing/ops_view.rs
@@ -239,7 +239,7 @@ async fn query_workspace_subscription(
             id,
             workspace_type,
             plan_tier,
-            daily_job_quota,
+            daily_job_quota::bigint,
             billing_provider,
             billing_customer_id,
             billing_currency,
